libstagefright: Remove channel count check during AAC configuration
- AAC format configuration in stagefright has CHECK to verify if channel
count value is not equal to 1 or 2, in which case it would result in
mediaserver crash.

- It is observed that on parsing the codec specific information of ESDS
atom in few AAC clips, mpeg4 extractor gives channel count value as
zero and results in crash because of above check. Though we will be
able to detect the correct channel count later from the AAC bitstream.

- OMX component will be able to detect the correct channel count value
and can reconfigure the port settings if the channel count value is
wrongly set and this check is not required in OMX client code of
stagefright. Hence removing this check to avoid mediaserver crash and
allow playing these clips which generate wrong channel count value.

Change-Id: Idb28d3805746191e8556bfba217a87f530327aa8
CRs-Fixed: 271979