libstagefright: Fix for handling empty buffers from extractors
There is a possiblity that mSource->read will return a NULL buffer with
a invalid size.  As some implementations of memcpy don't handle NULL
buffers well, it's best to check for a NULL buffer and avoid the memcpy.

This fix also handles the case where the buffer happens to be NULL but
size happens to be wrongly set.

(cherry picked from commit b05336ce6a579a2a4912177b0606d25430382abf)

Change-Id: Ic48ba92c7bc73afefc3b63e5cf00cb96ee384175

Conflicts:

	media/libstagefright/OMXCodec.cpp